# bpo-41513: Improve order of adding fractional values. Improve variable names.